Move constants to constants.js; move `generatePointerHash` to util.js
