+++ packages/dd-trace/test/util.spec.js
@@ -3,6 +3,7 @@
 require('./setup/tap')
 
 const { isTrue, isFalse, globMatch } = require('../src/util')
+const { generatePointerHash } = require('../src/util')
 
 const TRUES = [
   1,
@@ -66,5 +67,22 @@ describe('util', () => {
     NONMATCH_CASES.forEach(({ subject, pattern }) => {
       expect(globMatch(pattern, subject)).to.equal(false)
     })
+  })
+})
+
+describe('generatePointerHash', () => {
+  it('should generate a valid hash for a basic S3 object', () => {
+    const hash = generatePointerHash(['some-bucket', 'some-key.data', 'ab12ef34'])
+    expect(hash).to.equal('e721375466d4116ab551213fdea08413')
+  })
+
+  it('should generate a valid hash for an S3 object with a non-ascii key', () => {
+    const hash1 = generatePointerHash(['some-bucket', 'some-key.你好', 'ab12ef34'])
+    expect(hash1).to.equal('d1333a04b9928ab462b5c6cadfa401f4')
+  })
+
+  it('should generate a valid hash for multipart-uploaded S3 object', () => {
+    const hash1 = generatePointerHash(['some-bucket', 'some-key.data', 'ab12ef34-5'])
+    expect(hash1).to.equal('2b90dffc37ebc7bc610152c3dc72af9f')
   })
 })
